requirements.txt
groq
python-dotenv
markdown2
pdfkit
